requirements.txt
asgiref==3.7.1
django==4.2.1
django_extensions==3.2.1
djangorestframework==3.14.0
Pillow==9.5.0
urllib3==2.0.2
webdriver_manager==3.8.6
wsproto==1.2.0
selenium
python-dotenv